% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.eks_operations.R
\name{list_updates}
\alias{list_updates}
\title{Lists the updates associated with an Amazon EKS cluster in your AWS account, in the specified Region}
\usage{
list_updates(name, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{name}{[required] The name of the Amazon EKS cluster for which to list updates.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated \code{ListUpdates} request where \code{maxResults} was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the \code{nextToken} value.}

\item{maxResults}{The maximum number of update results returned by \code{ListUpdates} in paginated output. When this parameter is used, \code{ListUpdates} only returns \code{maxResults} results in a single page along with a \code{nextToken} response element. The remaining results of the initial request can be seen by sending another \code{ListUpdates} request with the returned \code{nextToken} value. This value can be between 1 and 100. If this parameter is not used, then \code{ListUpdates} returns up to 100 results and a \code{nextToken} value if applicable.}
}
\description{
Lists the updates associated with an Amazon EKS cluster in your AWS account, in the specified Region.
}
\section{Accepted Parameters}{
\preformatted{list_updates(
  name = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

