% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.eks_package.R
\docType{package}
\name{paws.eks-package}
\alias{paws.eks}
\alias{paws.eks-package}
\title{paws.eks: Amazon Elastic Container Service for Kubernetes}
\description{
Amazon Elastic Container Service for Kubernetes
    (Amazon EKS) is a managed service that makes it easy for you to run
    Kubernetes on AWS without needing to stand up or maintain your own
    Kubernetes control plane. Kubernetes is an open-source system for
    automating the deployment, scaling, and management of containerized
    applications.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
